refactor: move worker and mdw websocket to nest services, rename worker to PairSyncService
