+++ src/app.module.ts
@@ -5,10 +5,12 @@ import { TokensService } from './api/tokens/tokens.service';
 import { PairsService } from './api/pairs/pairs.service';
 import { ClientsModule } from './clients/clients.module';
 import { ApiModule } from './api/api.module';
+import { PairSyncService } from './tasks/pair-sync.service';
+import { MdwWsClientService } from './clients/mdw-ws-client.service';
 
 @Module({
   imports: [ApiModule, ClientsModule, DatabaseModule],
   controllers: [AppController],
-  providers: [TokensService, PairsService],
+  providers: [MdwWsClientService, PairsService, TokensService, PairSyncService],
 })
 export class AppModule {}
