+++ src/tasks/pair-liquidity-info-history-importer.service.spec.ts
@@ -1,11 +1,11 @@
 import { PairLiquidityInfoHistoryImporterService } from './pair-liquidity-info-history-importer.service';
 import { Test, TestingModule } from '@nestjs/testing';
-import { MdwClientService } from '../clients/mdw-client.service';
+import { MdwHttpClientService } from '../clients/mdw-http-client.service';
 import { PairDbService } from '../database/pair/pair-db.service';
 import { PairLiquidityInfoHistoryDbService } from '../database/pair-liquidity-info-history/pair-liquidity-info-history-db.service';
 import { PairLiquidityInfoHistoryErrorDbService } from '../database/pair-liquidity-info-history-error/pair-liquidity-info-history-error-db.service';
 import { ContractAddress } from '../lib/utils';
-import { Contract } from '../clients/mdw-client.model';
+import { Contract } from '../clients/mdw-http-client.model';
 
 const mockMdwClientService = {
   getContract: jest.fn(),
@@ -35,7 +35,7 @@ describe('PairLiquidityInfoHistoryImporterService', () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         PairLiquidityInfoHistoryImporterService,
-        { provide: MdwClientService, useValue: mockMdwClientService },
+        { provide: MdwHttpClientService, useValue: mockMdwClientService },
         { provide: PairDbService, useValue: mockPairDb },
         {
           provide: PairLiquidityInfoHistoryDbService,
