+++ src/tasks/tasks.service.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { TasksService } from './tasks.service';
 import { PairLiquidityInfoHistoryImporterService } from './pair-liquidity-info-history-importer.service';
 import { PairLiquidityInfoHistoryValidatorService } from './pair-liquidity-info-history-validator.service';
-import { MdwClientService } from '../clients/mdw-client.service';
+import { MdwHttpClientService } from '../clients/mdw-http-client.service';
 import { PairDbService } from '../database/pair/pair-db.service';
 import { PairLiquidityInfoHistoryDbService } from '../database/pair-liquidity-info-history/pair-liquidity-info-history-db.service';
 import { PairLiquidityInfoHistoryErrorDbService } from '../database/pair-liquidity-info-history-error/pair-liquidity-info-history-error-db.service';
@@ -19,7 +19,7 @@ describe('TasksService', () => {
         TasksService,
         PairLiquidityInfoHistoryImporterService,
         PairLiquidityInfoHistoryValidatorService,
-        MdwClientService,
+        MdwHttpClientService,
         PairDbService,
         PairLiquidityInfoHistoryDbService,
         PairLiquidityInfoHistoryErrorDbService,
